requirements_dev.txt
bottle

docutils
MarkupSafe
pkginfo
Pygments
Sphinx
sphinx-rtd-theme
cloud-sptheme
sphinxcontrib-applehelp
sphinxcontrib-devhelp
sphinxcontrib-htmlhelp
sphinxcontrib-jsmath
sphinxcontrib-qthelp
sphinxcontrib-serializinghtml
sphinxcontrib-fulltoc
sphinxjp.themes.basicstrap
sphinxcontrib-mermaid
sphinx-autobuild
sphinx-tabs
tqdm
twine

matplotlib
mypy
pillow

pytest
splinter[selenium]
pytest-splinter
pytest-cov
WebTest
